+++ tests/governance.py
@@ -1,7 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
 import os
-import sys
 import http
 import subprocess
 import infra.network
@@ -13,6 +12,7 @@ import infra.e2e_args
 import suite.test_requirements as reqs
 import infra.logging_app as app
 import json
+import requests
 
 from loguru import logger as LOG
 
@@ -20,6 +20,10 @@ from loguru import logger as LOG
 @reqs.description("Test quotes")
 @reqs.supports_methods("quotes/self", "quotes")
 def test_quote(network, args):
+    if args.enclave_type == "virtual":
+        LOG.warning("Quote test can only run in real enclaves, skipping")
+        return network
+
     primary, _ = network.find_nodes()
     with primary.client() as c:
         oed = subprocess.run(
@@ -200,6 +204,45 @@ def test_service_principals(network, args):
 def test_ack_state_digest_update(network, args):
     for node in network.get_joined_nodes():
         network.consortium.get_any_active_member().update_ack_state_digest(node)
+    return network
+
+
+@reqs.description("Test invalid client signatures")
+def test_invalid_client_signature(network, args):
+    primary, _ = network.find_primary()
+
+    def post_proposal_request_raw(node, headers=None, expected_error_msg=None):
+        r = requests.post(
+            f"https://{node.pubhost}:{node.pubport}/gov/proposals",
+            headers=headers,
+            verify=os.path.join(node.common_dir, "networkcert.pem"),
+        ).json()
+        assert r["error"]["code"] == "InvalidAuthenticationInfo"
+        assert (
+            expected_error_msg in r["error"]["message"]
+        ), f"Expected error message '{expected_error_msg}' not in '{r['error']['message']}'"
+
+    # Verify that _some_ HTTP signature parsing errors are communicated back to the client
+    post_proposal_request_raw(
+        primary,
+        headers=None,
+        expected_error_msg="Missing signature",
+    )
+    post_proposal_request_raw(
+        primary,
+        headers={"Authorization": "invalid"},
+        expected_error_msg="'authorization' header only contains one field",
+    )
+    post_proposal_request_raw(
+        primary,
+        headers={"Authorization": "invalid invalid"},
+        expected_error_msg="'authorization' scheme for signature should be 'Signature",
+    )
+    post_proposal_request_raw(
+        primary,
+        headers={"Authorization": "Signature invalid"},
+        expected_error_msg="Error verifying HTTP 'digest' header: Missing 'digest' header",
+    )
 
 
 def run(args):
@@ -215,13 +258,11 @@ def run(args):
         network = test_no_quote(network, args)
         network = test_service_principals(network, args)
         network = test_ack_state_digest_update(network, args)
+        network = test_invalid_client_signature(network, args)
 
 
 if __name__ == "__main__":
     args = infra.e2e_args.cli_args()
-    if args.enclave_type == "virtual":
-        LOG.warning("This test can only run in real enclaves, skipping")
-        sys.exit(0)
 
     args.package = "liblogging"
     args.nodes = infra.e2e_args.max_nodes(args, f=0)